Show verbose_name_plural to humans
Currently we show `view.plural_name` in the title attribute of the
Browse page. We also use that in other locations in libya_elections.

There are 2 problems with this.

1) `plural_name` is defined on the bread object, not the view object
2) `plural_name` is meant to be used in URLs, so it's not human-friendly.

Attached is a proposed solution: Include `model_meta` in the template and use `model_meta.verbose_name_plural`.

If this is a reasonable solution, then I will implement it in libya_elections as well.