More flexible template resolution
This implements the proposal suggested in #28.

Note that I updated the documentation to make it clear that the
app-specific templates need to be formatted `{model}{activity}.html`
without an underscore. E.g `smsread.html`, not `sms_read.html`. This is
because we have set `template_name_suffix` to be `activity.html`, not
`_activity.html` as they are in Django Vanilla Views.