58 Handle exception from related object does not exist
There are some cases in Django where referencing a model field
that is a link to another model that doesn't exist will raise a
RelatedObjectDoesNotExist exception. We were catching that in
our template tag, but logging an exception. Depending on the logging
configuration, this could result in administrators getting an error
email from the server.

Fix by silently ignoring these exceptions, and just returning None
(which is what you might have expected Django to do anyway).

Fixes #58