Use Django 2.1's "view_{model_name}" permission
Django 2.1 added a standard permission for read-only access,
and it's named "view_*" instead of the non-standard "read_*"
we've been using here. Change the default view permission on
the read-only view to need "view_*" instead of "read_*".
This could be a breaking change, so add a warning to the README
about it.